feat(nfts): handle pending eth for sale, cancel offer, and cancel listing
