+++ packages/blockchain-wallet-v4-frontend/src/modals/Nfts/NftOrder/AcceptOffer/index.tsx
@@ -1,16 +1,13 @@
 import React from 'react'
 import { FormattedMessage } from 'react-intl'
-import { connect, ConnectedProps, useDispatch } from 'react-redux'
+import { connect, ConnectedProps } from 'react-redux'
 
-import { Remote } from '@core'
-import { Button, HeartbeatLoader, Link, Text } from 'blockchain-info-components'
+import { Text } from 'blockchain-info-components'
 import CoinDisplay from 'components/Display/CoinDisplay'
 import FiatDisplay from 'components/Display/FiatDisplay'
 import { Flex } from 'components/Flex'
 import FlyoutHeader from 'components/Flyout/Header'
-import { actions } from 'data'
 import { RootState } from 'data/rootReducer'
-import { Analytics } from 'data/types'
 import { useRemote } from 'hooks'
 
 import { NftFlyoutRow, StickyCTA } from '../../components'
@@ -18,24 +15,15 @@ import NftAssetHeaderRow from '../../components/NftAssetHeader'
 import NftFlyoutFailure from '../../components/NftFlyoutFailure'
 import NftFlyoutLoader from '../../components/NftFlyoutLoader'
 import { Props as OwnProps } from '..'
+import AcceptOfferCTA from './cta'
 import AcceptOfferFees from './fees'
 import { getData } from './selectors'
 
 const AcceptOffer: React.FC<Props> = (props) => {
-  const { close, isInvited, nftActions, openSeaAssetR, orderFlow } = props
-  const dispatch = useDispatch()
-  const acceptOfferClicked = () => {
-    dispatch(
-      actions.analytics.trackEvent({
-        key: Analytics.NFT_ACCEPT_OFFER_CLICKED,
-        properties: {}
-      })
-    )
-  }
+  const { close, openSeaAssetR, orderFlow } = props
 
   const openSeaAsset = useRemote(() => openSeaAssetR)
 
-  const disabled = Remote.Loading.is(orderFlow.fees) || props.orderFlow.isSubmitting
   if (openSeaAsset.isLoading) return <NftFlyoutLoader close={props.close} />
   if (openSeaAsset.error)
     return <NftFlyoutFailure error={openSeaAsset.error || ''} close={props.close} />
@@ -86,60 +74,7 @@ const AcceptOffer: React.FC<Props> = (props) => {
         <StickyCTA>
           <AcceptOfferFees {...props} asset={asset} />
           <br />
-          {isInvited ? (
-            props.data.cata({
-              Failure: (e) => (
-                <>
-                  <Text
-                    size='14px'
-                    weight={600}
-                    style={{ marginBottom: '8px', maxHeight: '200px' }}
-                  >
-                    {e}
-                  </Text>
-                  <Button jumbo nature='sent' fullwidth data-e2e='n/a' disabled>
-                    <FormattedMessage id='copy.accept_offer' defaultMessage='Accept Offer' />
-                  </Button>
-                </>
-              ),
-              Loading: () => (
-                <Button jumbo nature='primary' fullwidth data-e2e='n/a' disabled>
-                  <FormattedMessage id='copy.accept_offer' defaultMessage='Accept Offer' />
-                </Button>
-              ),
-              NotAsked: () => null,
-              Success: (val) => (
-                <Button
-                  jumbo
-                  nature='primary'
-                  fullwidth
-                  data-e2e='acceptNftOffer'
-                  disabled={disabled}
-                  type='submit'
-                  onClick={() => {
-                    acceptOfferClicked()
-                    nftActions.acceptOffer({
-                      asset,
-                      gasData: val.fees,
-                      ...val.matchingOrder
-                    })
-                  }}
-                >
-                  {props.orderFlow.isSubmitting ? (
-                    <HeartbeatLoader color='blue100' height='20px' width='20px' />
-                  ) : (
-                    <FormattedMessage id='copy.accept_offer' defaultMessage='Accept Offer' />
-                  )}
-                </Button>
-              )
-            })
-          ) : (
-            <Link href='https://www.blockchain.com/waitlist/nft' target='_blank'>
-              <Button jumbo nature='primary' fullwidth data-e2e='joinWaitlist'>
-                <FormattedMessage id='copy.join_waitlist' defaultMessage='Join the Waitlist' />
-              </Button>
-            </Link>
-          )}
+          <AcceptOfferCTA {...props} asset={asset} />
         </StickyCTA>
       </div>
     </>
@@ -152,6 +87,6 @@ const mapStateToProps = (state: RootState) => ({
 
 const connector = connect(mapStateToProps)
 
-type Props = OwnProps & ConnectedProps<typeof connector>
+export type Props = OwnProps & ConnectedProps<typeof connector>
 
 export default connector(AcceptOffer)
