+++ packages/blockchain-wallet-v4-frontend/src/modals/Nfts/NftOrder/MarkForSale/index.tsx
@@ -11,7 +11,7 @@ import styled from 'styled-components'
 
 import { getRatesSelector } from '@core/redux/data/misc/selectors'
 import { RatesType } from '@core/types'
-import { Button, HeartbeatLoader, Icon, Link, Text } from 'blockchain-info-components'
+import { Icon, Link, Text } from 'blockchain-info-components'
 import { Flex } from 'components/Flex'
 import { Title, Value } from 'components/Flyout'
 import FlyoutHeader from 'components/Flyout/Header'
@@ -29,6 +29,7 @@ import NftAssetHeaderRow from '../../components/NftAssetHeader'
 import NftFlyoutFailure from '../../components/NftFlyoutFailure'
 import NftFlyoutLoader from '../../components/NftFlyoutLoader'
 import { Props as OwnProps } from '..'
+import MarkForSaleCTA from './cta'
 import MarkForSaleFees from './fees'
 import { NftMarkForSaleFormValues } from './MarkForSale.types'
 import { getQuoteAmts } from './MarkForSale.utils'
@@ -59,34 +60,11 @@ const SaleSelection = styled.div`
 `
 
 const MarkForSale: React.FC<Props> = (props) => {
-  const {
-    analyticsActions,
-    close,
-    formValues,
-    isInvited,
-    nftActions,
-    openSeaAssetR,
-    orderFlow,
-    rates
-  } = props
+  const { analyticsActions, close, formValues, openSeaAssetR, orderFlow, rates } = props
   const { fix } = formValues || { fix: 'CRYPTO' }
   const [saleType, setSaleType] = useState<'fixed-price' | 'timed-auction'>('fixed-price')
   const [isReservedChecked, setIsReserveChecked] = useState<boolean>(false)
   const [open, setOpen] = useState(true)
-  const disabled = (
-    saleType === 'fixed-price'
-      ? // Fixed Price
-        !formValues?.fixAmount
-      : // Dutch (Declining)
-        formValues?.timedAuctionType === 'decliningPrice'
-  )
-    ? !formValues?.starting || !formValues?.ending || formValues?.ending >= formValues?.starting
-    : // English (Ascending)
-      !formValues.starting ||
-      (formValues.reserve &&
-        formValues.reserve >= formValues.starting &&
-        formValues?.timedAuctionType === 'highestBidder') ||
-      props.orderFlow.isSubmitting
 
   const resetForms = () => {
     formValues.fixAmount = ''
@@ -136,12 +114,12 @@ const MarkForSale: React.FC<Props> = (props) => {
         Failure: (e) => <NftFlyoutFailure error={e} close={close} />,
         Loading: () => <NftFlyoutLoader close={props.close} />,
         NotAsked: () => <NftFlyoutLoader close={props.close} />,
-        Success: (val) => (
+        Success: (asset) => (
           <>
             <FlyoutHeader sticky data-e2e='wrapEthHeader' mode='back' onClick={() => close()}>
               Sell Item
             </FlyoutHeader>
-            <NftAssetHeaderRow asset={val} />
+            <NftAssetHeaderRow asset={asset} />
             <div
               style={{
                 display: 'flex',
@@ -482,97 +460,9 @@ const MarkForSale: React.FC<Props> = (props) => {
               </NftFlyoutRow>
             </div>
             <StickyCTA>
-              <MarkForSaleFees {...props} asset={val} />
+              <MarkForSaleFees {...props} asset={asset} />
               <br />
-              {isInvited ? (
-                props.orderFlow.fees.cata({
-                  Failure: () => (
-                    <Button jumbo nature='sent' fullwidth data-e2e='sellNft' disabled>
-                      <FormattedMessage id='copy.sell_item' defaultMessage='Sell Item' />
-                    </Button>
-                  ),
-                  Loading: () => (
-                    <Button jumbo nature='primary' fullwidth data-e2e='sellNft' disabled>
-                      <FormattedMessage id='copy.sell_item' defaultMessage='Sell Item' />
-                    </Button>
-                  ),
-                  NotAsked: () => null,
-                  Success: (fees) => (
-                    <Button
-                      jumbo
-                      nature='primary'
-                      fullwidth
-                      data-e2e='sellNft'
-                      disabled={disabled}
-                      onClick={() => {
-                        if (saleType === 'fixed-price') {
-                          nftActions.createSellOrder({
-                            asset: val,
-                            endPrice: null,
-                            expirationMinutes: formValues.expirationMinutes,
-                            gasData: fees,
-                            paymentTokenAddress: undefined,
-                            reservePrice: undefined,
-                            startPrice: Number(amount),
-                            waitForHighestBid: false
-                          })
-                          // English Auction
-                        } else if (
-                          saleType === 'timed-auction' &&
-                          formValues?.timedAuctionType === 'highestBidder'
-                        ) {
-                          nftActions.createSellOrder({
-                            asset: val,
-                            endPrice: null,
-                            expirationMinutes: formValues.expirationMinutes,
-                            gasData: fees,
-                            paymentTokenAddress: window.coins.WETH.coinfig.type.erc20Address,
-                            reservePrice: Number(formValues.reserve),
-                            startPrice: Number(formValues.starting),
-                            waitForHighestBid: true
-                          })
-                        }
-                        // Dutch Auction
-                        else {
-                          nftActions.createSellOrder({
-                            asset: val,
-                            endPrice: Number(formValues.ending),
-                            expirationMinutes: formValues.expirationMinutes,
-                            gasData: fees,
-                            paymentTokenAddress: undefined,
-                            reservePrice: undefined,
-                            startPrice: Number(formValues.starting),
-                            waitForHighestBid: false
-                          })
-                        }
-
-                        analyticsActions.trackEvent({
-                          key: Analytics.NFT_SELL_ITEM_CLICKED,
-                          properties: {
-                            amount: Number(amount),
-                            collection: val.collection.name,
-                            collection_id: val.token_id,
-                            selling_fees: Number(fees.totalFees),
-                            type: saleType === 'fixed-price' ? 'FIXED_PRICE' : 'TIME_AUCTION'
-                          }
-                        })
-                      }}
-                    >
-                      {props.orderFlow.isSubmitting ? (
-                        <HeartbeatLoader color='blue100' height='20px' width='20px' />
-                      ) : (
-                        <FormattedMessage id='copy.sell_item' defaultMessage='Sell Item' />
-                      )}
-                    </Button>
-                  )
-                })
-              ) : (
-                <Link href='https://www.blockchain.com/waitlist/nft' target='_blank'>
-                  <Button jumbo nature='primary' fullwidth data-e2e='joinWaitlist'>
-                    <FormattedMessage id='copy.join_waitlist' defaultMessage='Join the Waitlist' />
-                  </Button>
-                </Link>
-              )}
+              <MarkForSaleCTA {...props} asset={asset} amount={amount} saleType={saleType} />
             </StickyCTA>
           </>
         )
@@ -608,6 +498,6 @@ const enhance = compose(
   connector
 )
 
-type Props = OwnProps & ConnectedProps<typeof connector>
+export type Props = OwnProps & ConnectedProps<typeof connector>
 
 export default enhance(MarkForSale) as React.FC<OwnProps>
